Move core service installation into core-services.js for easier refactoring
