+++ src/inabox/amp-inabox-lite.js
@@ -24,11 +24,15 @@ import {
   installGlobalNavigationHandlerForDoc,
 } from '../service/navigation';
 import {Services} from '../services';
-import {adopt, installBuiltins, installRuntimeServices} from '../runtime';
+import {adopt} from '../runtime';
 import {cssText as ampDocCss} from '../../build/ampdoc.css';
 import {cssText as ampSharedCss} from '../../build/ampshared.css';
 import {fontStylesheetTimeout} from '../font-stylesheet-timeout';
 import {getMode} from '../mode';
+import {
+  installBuiltinElements,
+  installRuntimeServices,
+} from '../service/core-services';
 import {installDocService} from '../service/ampdoc-impl';
 import {installErrorReporting} from '../error';
 import {installIframeMessagingClient} from './inabox-iframe-messaging-client';
@@ -100,7 +104,7 @@ installStylesForDoc(
     registerIniLoadListener(ampdoc);
 
     // Builtins.
-    installBuiltins(self);
+    installBuiltinElements(self);
     adopt(self);
 
     // Pre-stub already known elements.
